refactor(worker): progress
