+++ src/Worker/AbstractWorker.php
@@ -8,10 +8,14 @@ use DateTimeImmutable;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use SchedulerBundle\Event\TaskExecutedEvent;
 use SchedulerBundle\Event\TaskExecutingEvent;
+use SchedulerBundle\Event\WorkerRestartedEvent;
+use SchedulerBundle\Exception\LogicException;
 use SchedulerBundle\Runner\RunnerInterface;
 use SchedulerBundle\SchedulerInterface;
 use SchedulerBundle\Task\TaskExecutionTrackerInterface;
 use SchedulerBundle\Task\TaskInterface;
+use SchedulerBundle\Task\TaskList;
+use SchedulerBundle\Task\TaskListInterface;
 use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\Lock\LockInterface;
 use Symfony\Component\Lock\PersistingStoreInterface;
@@ -25,9 +29,19 @@ abstract class AbstractWorker implements WorkerInterface
 {
     private SchedulerInterface $scheduler;
     private EventDispatcherInterface $eventDispatcher;
-    private ?PersistingStoreInterface $store;
+    private PersistingStoreInterface $store;
     private TaskExecutionTrackerInterface $tracker;
     protected array $options;
+    protected bool $isRunning;
+    protected bool $shouldStop = false;
+    protected ?TaskInterface $lastExecutedTask = null;
+
+    protected function getTasks($tasks): array
+    {
+        $tasks = 0 === count($tasks) ? $this->scheduler->getDueTasks() : $tasks;
+
+        return is_array($tasks) ? $tasks : iterator_to_array($tasks);
+    }
 
     protected function handleTask(RunnerInterface $runner, TaskInterface $task): void
     {
@@ -64,6 +78,25 @@ abstract class AbstractWorker implements WorkerInterface
         return (new DateTimeImmutable('now', $this->scheduler->getTimezone()))->diff($updatedNextExecutionDate)->s + $this->options['sleepDurationDelay'];
     }
 
+    protected function checkTaskState(TaskInterface $task): bool
+    {
+        if (TaskInterface::UNDEFINED === $task->getState()) {
+            throw new LogicException('The task state must be defined in order to be executed!');
+        }
+
+        if (in_array($task->getState(), [TaskInterface::PAUSED, TaskInterface::DISABLED], true)) {
+            $this->logger->info(sprintf('The following task "%s" is paused|disabled, consider enable it if it should be executed!', $task->getName()), [
+                'name' => $task->getName(),
+                'expression' => $task->getExpression(),
+                'state' => $task->getState(),
+            ]);
+
+            return false;
+        }
+
+        return true;
+    }
+
     protected function dispatch(Event $event): void
     {
         if (null === $this->eventDispatcher) {
@@ -71,6 +104,48 @@ abstract class AbstractWorker implements WorkerInterface
         }
 
         $this->eventDispatcher->dispatch($event);
+    }
+
+    public function restart(): void
+    {
+        $this->stop();
+        $this->isRunning = false;
+        $this->failedTasks = new TaskList();
+        $this->shouldStop = false;
+
+        $this->dispatch(new WorkerRestartedEvent($this));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function stop(): void
+    {
+        $this->shouldStop = true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isRunning(): bool
+    {
+        return $this->isRunning;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFailedTasks(): TaskListInterface
+    {
+        return $this->failedTasks;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getLastExecutedTask(): ?TaskInterface
+    {
+        return $this->lastExecutedTask;
     }
 
     /**
